+++ priv/repo/migrations/20201118170035_add_ars_and_organization_to_published_message_projection.exs
@@ -1,0 +1,11 @@
+defmodule DealogBackoffice.Repo.Migrations.AddArsAndOrganizationToPublishedMessageProjection do
+  use Ecto.Migration
+
+  def change do
+    alter table(:published_messages) do
+      add :ars, :string
+      add :organization, :string
+    end
+    create index(:published_messages, [:ars])
+  end
+end